move fnv hashing support into librustc_data_structures
